Revision 1
Removes most warnings from code, adds few properties of matrices and
class changes
