+++ src/ijordan/matrixonator/view/MatrixOverviewController.java
@@ -7,16 +7,13 @@ import org.controlsfx.dialog.Wizard.WizardPane;
 import org.controlsfx.dialog.Wizard.LinearFlow;
 
 import ijordan.matrixonator.MainApp;
-import ijordan.matrixonator.model.Matrix;
-import javafx.event.ActionEvent;
+import ijordan.matrixonator.model.*;
 import javafx.fxml.FXML;
 import javafx.scene.control.Alert;
 import javafx.scene.control.Alert.AlertType;
-import javafx.scene.control.Button;
 import javafx.scene.control.ButtonBar.ButtonData;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.Dialog;
-import javafx.scene.control.DialogPane;
 import javafx.scene.control.Label;
 import javafx.scene.control.TableColumn;
 import javafx.scene.control.TableView;
@@ -48,6 +45,10 @@ public class MatrixOverviewController {
 
 	// Reference to the main application.
 	private MainApp mainApp;
+	
+	//Matrix Classes we require (since Revision 2)
+	private SimpleMatrixArithmetic SMA = new SimpleMatrixArithmetic();
+	private MatrixSolver MS = new MatrixSolver();
 
 	/**
 	 * The constructor. The constructor is called before the initialise()
@@ -297,7 +298,7 @@ public class MatrixOverviewController {
 	 * @param matrix
 	 */
 	private void alertMatrixData(Matrix matrix) {
-		Dialog dialog = new Dialog();
+		Dialog<Object> dialog = new Dialog<Object>();
 		dialog.setTitle(matrix.getName());
 		dialog.setHeaderText("Showing the data associated with " + matrix.getName());
 		ButtonType closeButtonType = new ButtonType("Close", ButtonData.OK_DONE);
@@ -323,7 +324,7 @@ public class MatrixOverviewController {
 	private void handleCalculateRREF() {
 		int selectedIndex = matrixTable.getSelectionModel().getSelectedIndex();
 		if (selectedIndex >= 0) {
-			alertMatrixData(matrixTable.getSelectionModel().getSelectedItem().reducedEchelonForm());
+			alertMatrixData(MS.reducedEchelonForm(matrixTable.getSelectionModel().getSelectedItem()));
 		} else {
 			// Nothing is selected
 			Alert alert = new Alert(AlertType.ERROR);
